Create project structure
Created basic files and made a .gitignore file to not commit the large
node_modules file, instead the developers can run "npm install" with
the package.json file.

Also created 3 configs for development, production and a default one.
They use webpack-merge to extend over each other.

Added 2 scripts for the webpack-dev-server and for building with the
production config.